+++ src/Component/VideoUploader/Adapter/AwsS3AdapterInterface.php
@@ -1,9 +1,9 @@
-<?php namespace App\Component\VideoUploader\Adapter\GaufretteAwsS3;
+<?php namespace App\Component\VideoUploader\Adapter;
 
 use App\Component\Cloud\DigitalOceanInterface;
 use Aws\S3\S3ClientInterface;
 
-interface GaufretteAwsS3Interface
+interface AwsS3AdapterInterface
 {
     public function getBucket(): string;
     public function getDigitalOcean(): DigitalOceanInterface;