+++ src/Component/VideoUploader/Adapter/GaufretteAwsS3CoconutOutput.php
@@ -1,8 +1,0 @@
-<?php namespace App\Component\VideoUploader\Adapter;
-
-use App\Component\VideoUploader\Adapter\GaufretteAwsS3\GaufretteAwsS3Adapter;
-
-class GaufretteAwsS3CoconutOutput extends GaufretteAwsS3Adapter
-{
-    
-}