+++ themes/AdminPanel_VelzonChild/assets/js/pages/video-platform-settings-edit.js
@@ -6,4 +6,19 @@ $( function ()
         btnRemoveSelector: ".btnRemoveField",
         btnAddSelector:    ".btnAddField"
     });
+    
+    let transcodedVideoUrlsType = $( '.transcodedVideoUrlsType:checked' ).val();
+    alert( transcodedVideoUrlsType );
+    if( transcodedVideoUrlsType == 'cloud_signed' ) {
+        $( '#SignedUrlExpirationFormGroup' ).show();
+    }
+    
+    $( '.transcodedVideoUrlsType' ).on( 'change', function( e ) {
+        if ( this.value == 'cloud_signed' ) {
+            $( '#SignedUrlExpirationFormGroup' ).show();
+        }
+        else {
+            $( '#SignedUrlExpirationFormGroup' ).hide();
+        } 
+    });
 });